requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==4.2.20
django-ninja==1.4.0
pillow==11.1.0
psycopg2-binary==2.9.9
pydantic==2.11.1
pydantic_core==2.33.0
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.0
